--- src/14-router/04-CustomLink/style.css
@@ -1,14 +1,0 @@
-
-li,
-a {
-  text-decoration: none;
-}
-.seleted {
-  font-size: 30px;
-  background-color: greenyellow;
-  text-decoration: none;
-}
-
-.seleted::before {
-  content: '>'
-}
